functions separated in another file
functions of adding data to the database were separated from the index
