+++ src/screens/AdoptionForm/index.js
@@ -5,69 +5,10 @@ import TopBar from '../../components/atoms/TopBar';
 import InputForm from '../../components/atoms/inputForm';
 import QuestionForm from '../../components/atoms/QuestionForm';
 import ButtonForm from '../../components/atoms/ButtonForm';
-import firestore from '@react-native-firebase/firestore';
-import auth from '@react-native-firebase/auth';
+import functions from './functions';
 
 const question1 = ['Si', 'Aveces', 'Muy amenudo', 'Nunca'];
 const question2 = ['Casa', 'Patio', 'Calle', 'Nose'];
-
-const addForm = (
-  telefono,
-  correo,
-  ciudad,
-  mascotas,
-  horas,
-  pacifico,
-  lugar,
-  navigation,
-) => {
-  if (
-    telefono !== '' &&
-    correo !== '' &&
-    ciudad !== '' &&
-    mascotas !== '' &&
-    pacifico !== '' &&
-    horas !== '' &&
-    lugar !== ''
-  ) {
-    firestore()
-      .collection('adopciones')
-      .add({
-        telefono: telefono,
-        correo: correo,
-        ciudad: ciudad,
-        mascotas: mascotas,
-        horas: horas,
-        Pacifico: pacifico,
-        lugarMascota: lugar,
-        uidUsuario: auth().currentUser.uid,
-        nombreUsuario: auth().currentUser.displayName,
-      })
-      .then(_ => {
-        Alert.alert('Se ha enviado tu respuesta.', '', [
-          {text: 'ok', onPress: () => navigation.goBack()},
-        ]);
-      })
-      .catch(error => {
-        Alert.alert('Error', error, [{text: 'ok'}]);
-      });
-  } else {
-    Alert.alert('Por favor rellena los campos', '', [{text: 'ok'}]);
-  }
-};
-
-const functions = ({
-  tel,
-  correo,
-  ciudad,
-  mascotas,
-  horas,
-  pacifico,
-  lugar,
-  navigation,
-}) => {
-  addForm(tel, correo, ciudad, mascotas, horas, pacifico, lugar, navigation);
-};
 
 export default function AdoptionForm({navigation}) {
   const [tel, setTel] = useState('');
@@ -77,6 +18,7 @@ export default function AdoptionForm({navigation}) {
   const [horas, setHoras] = useState('');
   const [pacifico, setPacifico] = useState('');
   const [lugar, setLugar] = useState('');
+  const variables = {tel, correo, ciudad, mascotas, horas, pacifico, lugar};
   return (
     <View>
       <TopBar navigation={navigation} iconVisible={false} />
@@ -131,15 +73,8 @@ export default function AdoptionForm({navigation}) {
       <ButtonForm
         text={'Enviar Formulario'}
         onPress={() => {
-          functions({
-            tel,
-            correo,
-            ciudad,
-            mascotas,
-            horas,
-            pacifico,
-            lugar,
-            navigation,
+          functions.add({
+            variables,
           });
         }}
       />
